+++ internal/integration/tidb_test.go
@@ -822,7 +822,7 @@ create table atlas_types_sanity
 					},
 					{
 						Name: "tTimestampFraction",
-						Type: &schema.ColumnType{Type: &schema.TimeType{T: "timestamp", Precision: 6},
+						Type: &schema.ColumnType{Type: &schema.TimeType{T: "timestamp", Precision: intp(6)},
 							Raw: "timestamp(6)", Null: true},
 						Default: &schema.RawExpr{
 							X: "CURRENT_TIMESTAMP(6)",
@@ -843,7 +843,7 @@ create table atlas_types_sanity
 					},
 					{
 						Name: "tTimestampFractionOnUpdate",
-						Type: &schema.ColumnType{Type: &schema.TimeType{T: "timestamp", Precision: 6},
+						Type: &schema.ColumnType{Type: &schema.TimeType{T: "timestamp", Precision: intp(6)},
 							Raw: "timestamp(6)", Null: true},
 						Default: &schema.RawExpr{
 							X: "CURRENT_TIMESTAMP(6)",
@@ -871,7 +871,7 @@ create table atlas_types_sanity
 					},
 					{
 						Name: "tYear",
-						Type: &schema.ColumnType{Type: &schema.TimeType{T: "year", Precision: t.intByVersion(map[string]int{"mysql8": 0}, 4)},
+						Type: &schema.ColumnType{Type: &schema.TimeType{T: "year", Precision: intp(t.intByVersion(map[string]int{"mysql8": 0}, 4))},
 							Raw: t.valueByVersion(map[string]string{"mysql8": "year"}, "year(4) unsigned"), Null: true},
 					},
 					{
